backend/requirements.txt
fastapi
uvicorn
python-multipart
gitpython
google-genai
requests
python-dotenv
pinecone